Rename Window -> Scene
A Window concept will probably be better suited in the renderer
instead, and having this split seems like a cleaner decision, if we ever
decide to have offscreen rendering etc.

Additionally, move away from the idea of a root item. I think a better
idea is that the scene can have geometry that is set *externally* (e.g.
from the window's geometry), and that the items inside it have geometry
that is totally separate.

We probably still ultimately want a single root item when it comes to
event delivery, but let's see how far we get with this gone.