Split rendering into 'render' and 'draw' stages
It struck me as odd and limiting to be walking through non-visual nodes
while drawing onto the surface.

This is a two-stage renderer that walks the tree of nodes to reduce to a
list of drawables, which are the primitives implemented by the rendering
backend. Drawing is then done by simply iterating this list and drawing
in order.

At the moment nodes aren't copied, so it wouldn't be safe to separate
the two stages. I'm thinking that there are smart ways we can copy
primitive nodes that would allow the program to continue during drawing,
potentially from a different goroutine. That could really loosen the API
restrictions for applications.