Implement a simple Image item
An Image renders a texture. A texture is not a clearly defined type:
right now, we have a FileTexture which will load in from disk, but I
expect we will want at least a MemoryTexture, and perhaps a
NetworkTexture (or perhaps that should be merged into FileTexture
somehow, but I'd like the name for the file case to remain simple, as
that is I think the most common usecase).