Implement Touchable and Moveable interfaces
Also expand the concept of touch grabs a little more. Right now all
grabs are implicit: an item that is Touchable or Tappable will
implicitly take the grab if it hasn't been taken already.

In the future, only Tappable will do this, and we'll have some way to
explicitly Grab() for Touchable, I think.