Move animations to advance based on a shared frame time
This makes the cost of Get() essentially a no-op, which is an added
side-benefit, but the primary point is making sure that all animations
tick together at the same rate.

I'm a little unsure how the API for this feels, but I think it's
probably fine? And exposing a Windowable interface also means that nodes
will have easy access to the geometry etc in future.