Implement new scene renderer
SceneRenderer is a generic implementation of the node tree traversal,
event delivery, and drawable resolution pieces of rendering. It's
theoretically reusable between backends, and cleans up all of that logic
significantly.

The tree resolving algorithm is also improved to deliver events
properly and, hopefully, for better future performance. Briefly, the new
algorithm is to:

1) Resolve the entire tree, rendering nodes and gathering a list of
drawables and input handlers
2) Deliver all input events
3) If input events dirtied the scene, re-render to sync back up and
generate a new drawables list.

For now, the second pass is a full pass, but if we implement dirty
states for nodes that could be improved. The second pass will be skipped
if no input events are delivered, which is the common case, so it's not
a huge hit already.