Copy renderer to renderer2 for experimentation, and flesh out this idea some
The basic principle I'm following is that the renderer resolves the tree, keeping
a copy in SceneRenderer of the original instructions as it goes for
state preservation. When a NodeInstruction is a RenderableNode, that
node is resolved until it can't be rendered anymore (meaning: it must be
primitives like TransformNode/GeometryNode).

There's plenty missing here, I'm sure I'm going to miss stuff, but
here's the basics I think are most important:

* Right now, we handle RenderableNode top down. I don't think this works
  with layouts, but am not sure. It might be that this can be made to work
  if a layout resolves its children itself, but perhaps the existing
  renderer hooks are a better idea.
* State handling is nonexistent right now (apart from the basic concept
  of the renderer holding it, the RenderableNode allocating and
  manipulating it). What this needs to do is to pass a reference to the
  (renderer-held) variable, and let the RenderNodes populate/change it.
* Relatedly, It's an open question how state should work when descending down
  the node tree: it might be that RenderableNodes should actually create shadow
  nodes until actually reaching the drawable, unlike the current approach of just
  ignoring anything until it reaches a drawable.

Input handling is gone right now. I think might work in resolveTree once state
works: check the old node's state for any event handlers, and pass events in as
needed. Don't check the new tree to avoid having to re-render after the input
event has been processed. Though this means that there might be a frame of
latency before new items get input events, I think the reduction in complexity
and headache that rerendering would cause is worth it. Alternatively, we
could add an initial state allocation field to RenderableNode, but this
might mean more GC thrashing. Needs thought.

Rendering is also totally neutered right now. I think that using
Drawables may not make sense now we have a persistent shadow tree. I
don't have a clear vision on this, but some thoughts follow.

It might be worth splitting the monolithic render into a few different
goroutines, that communicate through channels

* main loop (gets input events from the windowing system)
* renderer/event delivery (runs user code)
* renderer (this may live on the main loop, it may not, it's kind of an
  undefined piece)

The idea is that the main loop does as little as possible. It passes a
render "job" off when it's time to prepare a new frame, containing the
latest snapshot of any events.

The renderer keeps track of the root node (or in future dirty state)
itself, and delivers the snapshot of event state to any
interested/eligable event handlers in the scene, as well as running
render down the tree and building a new shadow tree.

Once that shadow tree is complete, it passes the shadow tree back to the
renderer for it to go and do its thing on.

Idea:
  separate shadow trees, such that we have one front and back buffer
  (a shadow-shadow-tree) so that the frame preparation could run ahead of the
  render pass, but maybe this doesn't make sense due to input latency
  and it should just be a single tree.

Idea:
  with separate shadow trees (one for current, one for preparing)
  though, if we have some kind of dirty state and subtree rendering, we could also
  introduce the idea of urgency to that dirty state. for instance, an
  animation-driven dirty should always be given the highest priority so as
  to keep the screen moving nicely, for instance. this could be reused
  for things that have some form of built-in gesture control like a
  pannable viewport: it could force itself to move its content at a
  higher priority than the rest of the scene updates.

... there's a lot to unpack in terms of ideas and such here, and this
all requires a lot of thought and discussion, but hopefully it's some
good ideas on where this could end up going.