Make an internal interface for renderable node types
It shouldn't be necessary for user code to create new renderable node
types -- this is why we have props. That is not what this interface is
for.

API-wise, it's very limiting to only allow dynamic node trees via
RenderableNode. The immediate case is that I want to use a separate node
type with slightly different semantics for input events, but would like
it to behave exactly as a RenderableNode in other ways.

This one-function interface covers the renderer's needs and allows us to
make different node types that are ultimately renderables.

Now that I think about it, this would also enable some syntactic sugar
in the future, like:

type RectangleNode struct {
    RectangleProps
    SimpleParentable
}

func (n RectangleNode) Render(state *RenderState) sg.Node {
    return RectangleNodeRender(n.RectangleProps, state)
}

Which would be exactly equivalent to a RenderableNode with
RectangleNodeRender, except that it can be constructed as just:

RectangleNode{Geometry: geo}

Hmmmm.