Add a not-yet-functional TouchNode
TouchNode is a renderable node type that mirrors the RenderableNode API
except in that it provides a TouchRenderState rather than RenderState.
The TouchRenderState includes the state of touch events for this node,
including callbacks to be used for input events following this render.

The renderer side of input events is not implemented at all yet, so this
currently does nothing.