Demonstrate event delivery with hover events
SceneRender and InputHelper are now capable of delivering input events
to TouchNodes in the scene. For the moment, only hover state is
implemented.

The node-state-based touch is feeling extremely powerful so far. It's
also fantastic that InputHandler isn't actually dealing with nodes at
all -- it just gets geometry and a TouchState to act on.