Fix geometry in input handling
This code would transform x,y and w,h using the matrix, and treat the
latter as the transformed width and height. That is incorrect; the width
and height would include the translation, for example.

The accurate approach here is to transform the top-left (x,y) and
bottom-right (x+w,y+h), and use x2-x1,y2-y1 as w,h.