Use InputState to track event state for an InputNode
InputState is a new struct containing 1) all of the node-specific state
necessary for accurate event delivery, and 2) input parameters that are
useful to input event handlers, such as (eventually) mouse position.

InputNodes keep their state in the shadow tree in the same way as
RenderableNodes. It looked to me like there were some ideas about the
InputHelper tracking node-specific state; I think this would largely be
a mistake, since there is no way for it to track a node through changes
in the tree.

With this change, HoverableRect now behaves correctly (again), and
receives OnEnter/OnLeave only when these states actually change.