Add a more useful Geometry type
This logic was starting to get messy. Geometry is now a struct with
X,Y,Width,Height and useful functions for working with rectangles,
notably including TransformedBounds to get a bounding box of a
transformed geometry.

I deliberately chose to not base this off of Vec4, because the mechanics
of working with a width/height are very different from any generic Vec4
algorithm.