Implement mouse grab and press/release/move events
The problem to solve here was how to associate grabs (which are stateful
over the entire scene) with input nodes, which can disappear at any
time and can only reliably store state in the tree.

The solution here is for InputHelper to track the *InputState of the
current grab and flag whether it's seen in each pass of input events.
With a post-process function, that flag is checked and the grab is
cleaned up if the relevant input node is no longer part of the tree.

With this, simple press/release/move/grab works, although it definitely
needs to be exercised with some more complex cases.