Refactor resolveTree's child and render tree handling
The previous code turned into a fairly complex nest of logic, primarily
trying to handle reparenting of children to rendered nodes and
preserving state. It also, inconveniently, required that the rendered
tree be fully resolved before handling other children. That restriction
is about to become a problem for the box layout algorithm.

The new logic is functionally equivalent, except that children of
Components are reparented by pre-populating the rendered node's
shadowChildren with unresolved placeholders. When resolveTree is called
for the rendered node, those children are stacked appropriately and
handled in the same way as direct children.

That also allowed pulling apart the special-case rendered tree logic
from the parentable logic, which cleans it up substantially.