Remove parallel resolveTree from SceneRenderer
This hasn't shown any performance benefit in the real world so far, and
for most trees the costs probably exceed any possible benefit. It's
becoming complicated as the rendering algorithm grows, so I don't see
any reason to hold on to it for now.

I'm still interested in exploring parallelization in the renderer at
some point, but it will clearly need to be in more carefully targeted
places.