Require a Box at the root of the render tree
SceneRenderer will expect a Box at the root of the render tree,
primarily so that something is responsible for laying out any
first-level boxes in the scene.

This box should generally come from the window implementation and should
constrain the scene to the window geometry. That is now implemented for
the SDL renderer.