Remove nodes.Transform and apply transforms in Box
The Transform node has been removed; its functionality is now part of
Box, for good and valid reasons enumerated below. For the moment, Box
has a new Transform property that applies a transform to the Box and all
of it descendants. That API may change at some point.

There's another consequence of boxes and transforms: the effective
transform calculated at resolve time is no longer useful. Boxes position
their children after resolving the child tree, so the position can't be
factored into that transform without walking the tree again, which
becomes rather expensive at resolve time for deeply nested layouts.
I've gone for simply calculating transforms during walk; if caching ends
up being useful, we can add it back in later.

Transforms with box geometry should now be functional, though I expect
some edge cases (and especially layout cases) may not be correct yet.

From one perspective, this is moving against the goal of simple nodes
that do one thing. I'm choosing instead to stick to the principle of
'Box is the only source of geometry' strictly, because this has major
implications.

Box nodes interact with other Box nodes; for example, a Box layout is
laying out descendant Boxes, and isn't concerned with anything else.
This makes sense, because Box acts on geometry, and nothing else has
geometry. But that wasn't quite true: Transform nodes would also change
geometry (or the interpretation of it) for their children, and Transform
nodes could appear in two very inconvenient spots: between a Box and a
child Box, and between a Box and a drawable.

The Box->transform->drawable case is a pretty straightforward problem:
the drawable would apply the transform, but the layout had no idea about
it (because there is no Box that it applies to). That would result in
drawables violating the layout and constraints, unless you carefully put
a Box under any Transform.

The Box->transform->Box problem is what really motivated this solution.
In order to correctly present geometry to the parent and child boxes, we
would have to walk the tree to accumulate the relative transform between
parent and child. It was getting annoying to figure out -- but it's much
simpler when we know that transform cannot change other than from a Box.