Remove explicit size from nodes.Input
Input should be sized by boxes like everything else. The neat part here
is that because Input is not a Box, it doesn't _occupy_ any space on its
own, so it's safe to drop an Input in the midde of other layouts or
wherever it seems to belong. It can be wrapped in a Box to give it
specific geometry too, of course.