Add a basic Flex layout
Starting to try to stress test the Box design. This is the early (read:
easy) parts of the flexbox algorithm, and I'm happy with the results.

For the moment, this supports layout by row/column and reversed, basis
defined by pixels or percentage, and grow/shrink with appropriate
factors. It's also more-or-less coherent with the box consttraints
system, although there is room for improvement on that.

Code also exists to do an auto/content-defined basis, but it's currently
disabled because re-renders during layout don't work yet.