Allow SDL textures to be provided as an Image texture
This adds a CreateSDLTexture method to sdlsoftware.Window, which returns
the *sdl.Texture to user code. drawImage also implements drawing for
*sdl.Texture now, so it's possible for user code to provide that texture
in a nodes.Image directly.

This is SDL-specific, but it's a useful shortcut for efficient or more
flexible drawing. We still need more generic texture types at some
point.